refactor: move proxy to components
