+++ packages/http/src/index.ts
@@ -2,6 +2,7 @@ import { factory } from '@stoplight/prism-core';
 import { IHttpOperation } from '@stoplight/types';
 import { defaults } from 'lodash';
 import mock from './mocker';
+import { proxy } from './proxy';
 import route from './router';
 import { validateInput, validateOutput } from './validator';
 export * from './types';
@@ -16,5 +17,5 @@ export const createInstance = (
 ) =>
   factory<IHttpOperation, IHttpRequest, IHttpResponse, IHttpConfig>(
     defaultConfig,
-    defaults(components, { route, validateInput, validateOutput, mock }),
+    defaults(components, { route, validateInput, validateOutput, mock, proxy }),
   );
